Synced darwin and freebsd file accessor with the linux one.
Workaround upstream go library bug needs to be applied to darwin too
and probably freebsd as well.
Ref: https://github.com/golang/go/issues/27416
