+++ go.mod
@@ -129,7 +129,7 @@ require (
 	www.velocidex.com/golang/oleparse v0.0.0-20211013063943-0334d69593c1
 	www.velocidex.com/golang/regparser v0.0.0-20190625082115-b02dc43c2500
 	www.velocidex.com/golang/vfilter v0.0.0-20211116040728-425e2c8d4115
-	www.velocidex.com/golang/vtypes v0.0.0-20211201133246-fbb2618339a0
+	www.velocidex.com/golang/vtypes v0.0.0-20211203065440-c0c4ffabf7ad
 )
 
 require (
