--- utils/file_generic.go
@@ -1,20 +1,0 @@
-// +build darwin freebsd
-
-package utils
-
-import (
-	"io/ioutil"
-	"os"
-)
-
-func ReadDir(dirname string) ([]os.FileInfo, error) {
-	return ioutil.ReadDir(dirname)
-}
-
-func ReadDirUnsorted(dirname string) ([]os.FileInfo, error) {
-	return ioutil.ReadDir(dirname)
-}
-
-func ExpandEnv(path string) string {
-	return os.ExpandEnv(path)
-}
