+++ utils/stub.go
@@ -1,4 +1,4 @@
-// +build !linux
+// +build !linux,!darwin,!freebsd
 
 package utils
